Fix resolvesUponReachingToValue to Actually Work
I was erroneously assuming that ignoring velocity was sufficient to stop upon reaching the `toValue`, however with enough velocity, the check for whether `toValue` has been reached or not is incorrect.

Now, it checks to see if the delta between the current value and the value after executing the spring to see if there’s a sign change (i.e. crossing zero), meaning the `toValue` has been reached and the spring should resolve.