Support toValue on DecayAnimation; Add roundingFactor
This adds the ability to get the destination for a `DecayAnimation`. It also allows the ability to set the destination for a `DecayAnimation` (similar to how one can with `UIScrollView`). Setting the `toValue` on a `DecayAnimation` will result in the `velocity` being adjusted so that the `DecayAnimation` will stop at the supplied `toValue`.

This also adds the ability to adjust how the decay will end with regards to rounded values. Setting a value of 1.0 will result in the `DecayAnimation` always stopping on whole point values (whereas setting something like `1 / screenScale` will allow it to stop on pixel values).

Implements #28